modules/functions/requirements.txt
google-cloud-storage==2.10.0
